+++ src/scenes/Level.ts
@@ -931,6 +931,7 @@ export default class Level extends Phaser.Scene {
 		// velocity of knockback
 
 			this.player.setVelocityY(-this.player.jumpForce);
+			this.player.stateController.setState('airborne');
 
 			if (_enemy.hasParasol)
 			{
@@ -1849,7 +1850,7 @@ export default class Level extends Phaser.Scene {
 		// });
 			// TODO: have plane enemy position reflect this.
 
-		this.planeRect = new Phaser.Geom.Rectangle(x - 5, y - 15, this.plane.width, this.plane.height + 10);
+		this.planeRect = new Phaser.Geom.Rectangle(x - 5, y - 25, this.plane.width, this.plane.height + 20);
 		
 		if (__DEV__)
 		{